fixed floating AREF setting in lightdrv (it's very bad to do)
