+++ src/drive/lightdrv.cpp
@@ -3,20 +3,19 @@
 //
 
 #include "lightdrv.h"
+#include "../uart/uart.h"
 #include "../rtos/tta.h"
 
 static TaskHandle lightDrvTaskHandle = nullptr;
 
 struct lightDrvReading {
-    bool hit;
+    bool hit = false;
+    int rawVal = 0;
 };
 
 void _lightDrvTask(void * state){
     uint8_t high, low;
-    uint16_t result;
-
-    // raise MUX5 if reading from channel A8 or above, clear it if reading from channels A0-A7
-    ADCSRB = (ADCSRB & ~(1 << MUX5)) | (((LIGHTDRV_PIN >> 3) & 0x01) << MUX5);
+    int result;
 
     // select channel (mod 8, as high bit is handled by MUX5)
     ADMUX |= (LIGHTDRV_PIN & 0x07);
@@ -24,7 +23,6 @@ void _lightDrvTask(void * state){
     // start comparison by setting ADSC
     ADCSRA |= (1 << ADSC);
 
-    // TODO: is a poll here OK? I have it on word-of-forum that ADC takes maximum 15 cycles to complete; watch for trouble
     // wait for ADSC to clear on conversion finish
     while(ADCSRA & (1 << ADSC));
 
@@ -32,6 +30,12 @@ void _lightDrvTask(void * state){
     high = ADCH;
     result = (high << 8) | low;
 
+    char rawLevelStr[7];
+    utos(result, rawLevelStr);
+    uart_sendstr(rawLevelStr);
+    uart_sendstr("\t");
+
+    ((struct lightDrvReading *) state)->rawVal = 42;
     ((struct lightDrvReading *) state)->hit = (result > LIGHTDRV_COMPARE);
 }
 
@@ -39,8 +43,11 @@ void lightDriverStart(){
     // set pin data direction to 0 (input)
     LIGHTDRV_DDR &= ~(1<<LIGHTDRV_PIN);
 
-    // enable ADC comparisons
-    ADCSRA &= (1 << ADEN);
+    // use Vcc as reference
+    ADMUX = (1 << REFS0);
+
+    // 128x prescale clock, we are in no rush, and enable ADC
+    ADCSRA |= ( 1 << ADPS2 ) | ( 1 << ADPS1 ) | ( 1 << ADPS0 ) | (1 << ADEN);
 
     // start only one instance of LIGHTDRV
     if(!lightDrvTaskHandle)
@@ -53,5 +60,15 @@ bool readLight(){
         return ((struct lightDrvReading *) handle->state)->hit;
 
     //TODO: raise application-level error if we can't retrieve the LIGHTDRV state handle
-    return true;
+    return false;
+}
+
+int readLightLevel(){
+    StateHandle handle = OS_GetTaskState(lightDrvTaskHandle);
+    if(handle)
+        return ((struct lightDrvReading *) handle->state)->rawVal;
+
+    //TODO: raise application-level error if we can't retrieve the LIGHTDRV state handle
+    return -1;
+
 }
