+++ tests/build_test.sh
@@ -2,7 +2,9 @@
 rm *.o
 
 pname="${1%.*}"
-avr-g++ -c -std=c++0x -Os -mmcu=atmega2560 -Wa,--gstabs -o tta.o ../src/tta.cpp
+avr-g++ -c -std=c++0x -Os -mmcu=atmega2560 -Wa,--gstabs -o tta.o ../src/rtos/tta.cpp
+avr-g++ -c -std=c++0x -Os -mmcu=atmega2560 -Wa,--gstabs -o uart.o ../src/uart/uart.cpp
+avr-g++ -c -std=c++0x -Os -mmcu=atmega2560 -Wa,--gstabs -o lightdrv.o ../src/drive/lightdrv.cpp
 avr-g++ -c -std=c++0x -Os -mmcu=atmega2560 -Wa,--gstabs -o "$pname".o "$pname".cpp
 avr-g++  -mmcu=atmega2560 -o "$pname".elf *.o
 avr-objcopy -j .text -j .data -O ihex "$pname".elf "$pname".flash.hex
